requirements.txt
# Snapshot generator and historical data scripts
alpaca-py
psycopg2-binary
python-dotenv
pytz
requests
